Remove backups, copies, duplicates, etc.
Version control allows all of the changes made to files to be saved,
so there is no need to manually create copies, duplicates or versions
of files - such is the benefit of version control. In addition to greatly
simplifying the project tree and making it easier to understand whats
going on, it is less error prone. It is easier to understand what
files are supposed to do (and which files do it) when there is only
one 'version' of each file in the project.